Overall Version Update
Clips4Sale is working, but under construction to save c4s data to clipnuke.
PornHub is not working. Even added a prefill url so you can add ?clipnuke-id=XXXXXX to the new video page URL and it'll prefill the form with data from clipnuke for that video's ID. Works on Clips4Sale too.